+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 * Instance: Suppress action menu and disable buttons when click Change log icon. Refs UIIN-3176.
 * Holdings: Suppress action menu and disable buttons when click Change log icon. Refs UIIN-3177.
 * Item: Suppress action menu and disable buttons when click Change log icon. Refs UIIN-3178.
+* Use "number generator" for barcode, accession number and call number. Refs UIIN-2546.
 
 ## [12.0.12](https://github.com/folio-org/ui-inventory/tree/v12.0.12) (2025-01-27)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v12.0.11...v12.0.12)